Move file to correct path
